cmd, core: add encrypted API request IDs
Every /v1 API response now includes a Request-Id header with a unique ID
for the request. The same ID is added to request-bound logs, making it
easier to trace errors back to the API request that produced them.

Request-Id is returned for normal API responses and API error responses
once the request has entered the /v1 API server. It is not added to
paths outside /v1/.

In addition, clients can send a previous Request-Id back in
After-Request-Id:

   After-Request-Id: <request-id>

When Krenalis runs on multiple nodes, After-Request-Id lets clients
ensure that a later request sees the same state as an earlier request,
or a newer state.

Closes #193

+++ core/internal/state/notifier_test.go
@@ -12,7 +12,7 @@ func TestParsePayload(t *testing.T) {
 
 	tests := []struct {
 		notification string
-		id           int64
+		id           int
 		name         string
 		payload      string
 		err          bool
